refactor: page lookup into service
